fix plugin and app configuration specially in release build

+++ orthos/runtime/src/main/java/dev/igordesouza/orthos/runtime/signal/impl/NativeAgreement.kt
@@ -11,5 +11,5 @@ internal object NativeAgreement {
         System.loadLibrary("orthos")
     }
 
-    external fun nativeValue(): Int
+    external fun nativeValue(): Long
 }
